LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/unloading-documents/beforeunload-canceling-expected.txt

CONFIRM NAVIGATION: 
CONFIRM NAVIGATION: false
CONFIRM NAVIGATION: true
CONFIRM NAVIGATION: 0
CONFIRM NAVIGATION: foo
CONFIRM NAVIGATION: foo
CONFIRM NAVIGATION: foo
CONFIRM NAVIGATION: foo
CONFIRM NAVIGATION: foo
CONFIRM NAVIGATION: foo

PASS Returning a string must not cancel the event: CustomEvent, non-cancelable 
PASS Returning a string must not cancel the event: CustomEvent, cancelable 
PASS Returning null with a real iframe unloading 
PASS Returning undefined with a real iframe unloading 
PASS Returning  with a real iframe unloading 
PASS Returning false with a real iframe unloading 
PASS Returning true with a real iframe unloading 
PASS Returning 0 with a real iframe unloading 
PASS Returning null with a real iframe unloading; setting returnValue to foo 
PASS Returning undefined with a real iframe unloading; setting returnValue to foo 
PASS Returning  with a real iframe unloading; setting returnValue to foo 
PASS Returning false with a real iframe unloading; setting returnValue to foo 
PASS Returning true with a real iframe unloading; setting returnValue to foo 
PASS Returning 0 with a real iframe unloading; setting returnValue to foo 
 
